pyproject.toml
[tool.poetry]
name = "pydock"
version = "1.0.0"
description = "Python Docker Deployment Manager with REST API, interactive shell, and Cloudflare integration"
authors = ["PyDock Team <team@pydock.dev>"]
readme = "README.md"
homepage = "https://github.com/pydock/pydock"
repository = "https://github.com/pydock/pydock"
documentation = "https://pydock.readthedocs.io"
keywords = ["docker", "deployment", "vps", "automation", "devops", "cloudflare"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Systems Administration",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
paramiko = "^3.3.1"
pyyaml = "^6.0.1"
requests = "^2.31.0"
click = "^8.1.7"
colorama = "^0.4.6"
jinja2 = "^3.1.2"
cryptography = "^41.0.4"
python-dotenv = "^1.0.0"
rich = "^13.6.0"
typer = "^0.9.0"
httpx = "^0.25.0"
asyncio-mqtt = "^0.13.0"
pydantic = "^2.4.2"
pydantic-settings = "^2.0.3"
websockets = "^12.0"
python-multipart = "^0.0.6"
cloudflare = "^2.11.1"
GitPython = "^3.1.40"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
black = "^23.0.0"
flake8 = "^6.0.0"
mypy = "^1.5.0"
pre-commit = "^3.4.0"
pytest-cov = "^4.1.0"
mkdocs = "^1.5.0"
mkdocs-material = "^9.2.0"

[tool.poetry.scripts]
pydock = "pydock.cli:main"
pydock-server = "pydock.api.server:run_server"
pydock-shell = "pydock.shell:interactive_shell"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=pydock --cov-report=html --cov-report=term-missing"

[tool.coverage.run]
source = ["pydock"]
omit = ["tests/*", "pydock/templates/*"]